requirements.txt
streamlit==1.31.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
mypy==1.7.1